Craft the three crystal tiers, and only those (ticket 03)
The recipes are datapack JSON, so their ingredients are component-blind: a plain
echo shard in the centre matches the tier-2 pattern just as a tier-1 crystal
does. A guard at the crafting grid's own recompute says what the recipes cannot
— an upgrade must consume the tier below it, and a crystal is never an
ingredient in anyone else's recipe (the recovery compass being the case that
bites).

A unit test decodes the shipped JSON the way the datapack reload does and holds
it to CrystalItem.of(), because the component set now lives in two places. It
caught the first drift immediately: JSON has no integer widths, so the tier
narrows to a byte, and the builder now writes one too.

+++ src/main/kotlin/eu/mctraveler/crystal/CrystalItem.kt
@@ -54,7 +54,11 @@ object CrystalItem {
             crystal,
             CompoundTag().apply {
                 putBoolean(MARKER, true)
-                putInt(TIER, tier)
+                // A byte, not an int, so a crystal built here is byte-identical
+                // to one the datapack recipes craft: JSON has no integer widths,
+                // so the decoder narrows every small number to a byte. Reading
+                // goes through getIntOr, which takes any numeric tag either way.
+                putByte(TIER, tier.toByte())
             },
         )
         crystal.set(DataComponents.ITEM_NAME, Component.literal(ITEM_NAME))
